refactor: 프로젝트 structure를 변경하라
- 폴더 ui => components로 변경
- 폴더 model => types로 변경
