+++ app/about/page.tsx
@@ -1,9 +1,9 @@
 import { useMemo } from 'react';
 
 import ISO_3166_COUNTRY_CODES from 'lib/assets/data/iso3166CountryCodes';
-import { Country } from 'lib/model/country';
-import SearchCountry from 'ui/about/SearchCountry';
+import { Country } from 'lib/types/country';
 
+import SearchCountry from 'components/about/SearchCountry';
 import { codeToFlag } from 'utils';
 
 export const metadata = {
